+++ tempodb/encoding/common/interfaces.go
@@ -16,7 +16,7 @@ type Finder interface {
 	FindTraceByID(ctx context.Context, id ID, opts SearchOptions) (*tempopb.Trace, error)
 }
 
-type TagCallback func(t string)
+type TagCallback func(t string) (stop bool)
 
 type TagCallbackV2 func(traceql.Static) (stop bool)