+++ tempodb/encoding/vparquet/wal_block.go
@@ -539,8 +539,7 @@ func (b *walBlock) SearchTagValues(ctx context.Context, tag string, cb common.Ta
 
 	// Wrap to v2-style
 	cb2 := func(v traceql.Static) bool {
-		cb(v.EncodeToString(false))
-		return false
+		return cb(v.EncodeToString(false))
 	}
 
 	return b.SearchTagValuesV2(ctx, att, cb2, opts)