+++ tempodb/search/searchable_block.go
@@ -6,9 +6,12 @@ import (
 
 type tagCallback func(t string)
 
+// tagValueCallback is a callback for tag values.  If it returns true, the search will stop.
+type tagValueCallback func(v string) bool
+
 type SearchableBlock interface {
 	Tags(ctx context.Context, cb tagCallback) error
-	TagValues(ctx context.Context, tagName string, cb tagCallback) error
+	TagValues(ctx context.Context, tagName string, cb tagValueCallback) error
 	Search(ctx context.Context, p Pipeline, sr *Results) error
 }